pyproject.toml
[tool.poetry]
name = "yt-classifier"
version = "0.1.0"
description = ""
authors = ["Sungwoo Park <codest99@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.9.1"
google-api-python-client = "^2.86.0"
python-dotenv = "^1.0.0"
jupyterlab = "^4.0.0"
pandas = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
